fix: MaterialStatus removed as a parameter and added @Service for MachineQueryService.

+++ src/main/java/com/construtech/buildsphere/platform/resourceManagement/interfaces/rest/resources/CreateMachineResource.java
@@ -6,6 +6,6 @@ public record CreateMachineResource(
         String receptionDate,
         String endDate,
         double totalCost,
-        Long projectId
+        int projectId
 ) {
 }
